Use 'channel' parameter from config
This commit is for fetching the channel option from an optional
parameter in the bundle config.

If the param is not present the default value '32' will be used, because
the documentation specifies that if the param is not present in the
request all the payment options will be available.

See:
http://www.przelewy24.pl/eng/storage/app/media/pobierz/Instalacja/przelewy24_specification.pdf